[julia] put Julia code into a project better precompile
