+++ julia/Related/Project.toml
@@ -1,0 +1,10 @@
+name = "Related"
+uuid = "b1657308-8f9a-4fd8-ae9f-d732ff8d2de9"
+authors = []
+version = "0.1.0"
+
+[deps]
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
